+++ packages/add-months/index.ts
@@ -2,7 +2,7 @@
  * Increment or decrement the given date by `count` months.
  *
  * @param {Date} date Local date
- * @param {number} count Months to add or substract
+ * @param {number} count Months to add or subtract
  * @returns {Date} New local date
  * @private
  */
@@ -33,7 +33,7 @@ const addMonthsLoc = (date: Date, count: number) => {
  * Like the `addMonthsLoc` function, but using UTC methods.
  *
  * @param {Date} date UTC date
- * @param {number} count Months to add or substract
+ * @param {number} count Months to add or subtract
  * @returns {Date} New UTC date
  * @private
  */
@@ -75,24 +75,24 @@ const toDate = (src: any) => {
  * Returns a date occurring `count` months after `startdate` or, if `count` is
  * negative, the date occurring `count` months before `startdate`.
  *
- * - If startdate is not a Date, string, or number that can be converted to a
+ * - If `startdate` is not a Date or number that can be converted to a
  *    valid date, returns a new Date instance with an invalid date.
  *
- * - If count is evaluated as zero, returns a new Date instance with the
- *    the same value as startdate.
+ * - If `count` is evaluated as zero, returns a new Date instance with the
+ *    the same value as `startdate`.
  *
  * - If there is an overflow in the day, the date is adjusted to the last
  *    valid day of the expected month.
  *
  * The third parameter is optional and indicates if the date is UTC. It is
- * necessary to differentiate UTC dates from locals to avoid errors due to the
+ * necessary to differentiate UTC dates from locals and avoid errors due to the
  * [Daylight Saving Time](https://en.wikipedia.org/wiki/Daylight_saving_time)
  * (DST).
  *
  * This function does not change the original date.
  *
  * @param {Date|number} startdate A value parseable as a JavaScript Date
- * @param {number} count Number of months to add or substract
+ * @param {number} count Number of months to add or subtract
  * @param {boolean} [asUTC=false] If `true`, handle the date as UTC
  * @returns {Date} A new, adjusted Date instance.
  * @since 1.0.0