+++ packages/has-own-properties/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - Keywords to package.json
 
+### Changed
+
+- Separate logic, to enhance performance.
+
 ### Fixed
 
 - Issues with codebeat linters.