+++ packages/has-own-properties/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog for @jsbits/has-own-properties
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+For global changes, see the [changelog of jsbits](https://github.com/ProJSLib/jsbits/blob/master/CHANGELOG.md)
+
+## \[1.1.1] - 2018-12-29
+
+### Changed
+
+- Replace BSD-2 with the MIT license.
+- Update badges.
 
 ## \[1.1.0] - 2018-12-06